requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
Flask-Login==0.5.0
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.15.1
python-dotenv==0.19.0
psycopg2cffi==2.9.0
WTForms==2.3.3
Werkzeug==2.0.1